+++ src/ezdxf/__init__.py
@@ -14,11 +14,11 @@ PYPY_ON_WINDOWS = sys.platform.startswith('win') and PYPY
 from ezdxf.options import options
 from ezdxf.tools import transparency2float, float2transparency
 from ezdxf.tools.rgb import int2rgb, rgb2int
-from ezdxf.tools.pattern import PATTERN
 from ezdxf.lldxf import const
 from ezdxf.lldxf.validator import is_dxf_file, is_dxf_stream
 from ezdxf.filemanagement import readzip, new, read, readfile
 from ezdxf.tools.standards import setup_linetypes, setup_styles, setup_dimstyles, setup_dimstyle
+from ezdxf.tools import pattern
 from ezdxf.render.arrows import ARROWS
 from ezdxf.lldxf.const import DXFError
 from ezdxf.lldxf.const import DXFStructureError, DXFVersionError, DXFTableEntryError, DXFAppDataError, DXFXDataError