+++ src/ezdxf/options.py
@@ -13,6 +13,11 @@ class Options:
         self.default_text_style = 'OpenSans'
         self.default_dimension_text_style = 'OpenSansCondensed-Light'
 
+        # Predefined hatch pattern in ezdxf before v0.11 were scaled too small, by a factor of 1/25.4, compared with
+        # the predefined pattern in BricsCAD and AutoCAD, this was corrected in v0.11, but if your application
+        # depends on the old scaling and changing your code is too much work, set 'use_old_predefined_pattern_scaling'
+        # option to True to use the old pattern scaling.
+        self.use_old_predefined_pattern_scaling = False
         # debugging
         self.log_unprocessed_tags = True